Small fixes
