+++ babel.config.js
@@ -13,6 +13,8 @@ module.exports = function(api) {
             'language/*': './src/language/*',
             'settings': './src/components/settings',
             'settings/*': './src/components/settings/*',
+            'utils': './src/utils',
+            'utils/*': './src/utils/*',
           }
         }
       ]
