+++ tsconfig.json
@@ -16,6 +16,8 @@
       "language/*": ["src/language/*"],
       "settings": ["src/components/settings"],
       "settings/*": ["src/components/settings/*"],
+      "utils": ["src/utils"],
+      "utils/*": ["src/utils/*"],
     }
   },
   "include": ["App.tsx", "src/**/*"],
